Add relationship_path to agg and direct features
This adds an optional constructor parameter, and stores the
relationship_path on the feature. If no relationship_path is passed to
the constructor we search for a path between the target entity and base
entity. If none or more than one is found we raise an error.  When there
are more than one possible paths between entities the feature name will
include the full path.

Also add parent_name and child_name to Relationship.

This does not update any other logic to use the paths, as this will be
added in later commits.